mawk-patch-3 - patch3 for mawk1.1, Patch03
 Changes are:

 1) New configurations:
    hpux coherent 386bsd linux

 2) Changed some ints to longs so that Unix and Dos versions
    give same output.  E.g.,

    printf "%d" and printf "%ld" are now the same on Dos.

    msdos/examples directory with scripts by Ben Myers

    Infinity+1 large model pointer bug caused by modulo 2^16
    arithmetic fixed.

 3) Grossly surprised to discover  (int) d causes a floating
    point exception on hpux when d is a large double.
    Integer casts are now done more carefully.  Thanks to
    Ken Poulton for finding this bug.

 4) Removed ApolloSR10.3 from working configurations. Mawk will
    not work if compiled with CC6.8.

 5) A few minor fixes and tweaks.  ($Logs$ contain more detail).

+++ symtype.h
@@ -11,6 +11,9 @@ the GNU General Public License, version 2, 1991.
 ********************************************/
 
 /*$Log: symtype.h,v $
+ * Revision 5.3  1992/12/17  02:48:01  mike
+ * 1.1.2d changes for DOS
+ *
  * Revision 5.2  1992/07/08  15:44:44  brennan
  * patch2: length returns.  I am a wimp
  *
@@ -43,7 +46,7 @@ unsigned char min_args, max_args ;
 typedef  struct anode {
 struct anode *link , *ilink ;
 STRING *sval ;
-int     ival ;
+long     ival ;
 CELL   *cp ;
 }  ANODE ;
